tabletserver: Move infrastructure for /<table>z pages into an extra package go/vt/logz.
This way, other packages can reuse the code without depending on go/vt/tabletserver.

I'll use the new package in go/vt/throttler shortly.
